Split the app model into file open, filter, sync, and backup mixins.
