+++ app/scripts/models/app-model.js
@@ -1,6 +1,4 @@
 import { Events } from 'framework/events';
-import { Storage } from 'storage';
-import { SearchResultCollection } from 'collections/search-result-collection';
 import { FileCollection } from 'collections/file-collection';
 import { FileInfoCollection } from 'collections/file-info-collection';
 import { RuntimeInfo } from 'const/runtime-info';
@@ -9,24 +7,18 @@ import { TouchIdRecovery } from 'comp/app/touch-id-recovery';
 import { NativeModules } from 'comp/launcher/native-modules';
 import { Timeouts } from 'const/timeouts';
 import { AppSettingsModel } from 'models/app-settings-model';
-import { EntryModel } from 'models/entry-model';
 import { FileInfoModel } from 'models/file-info-model';
-import { FileModel } from 'models/file-model';
-import { GroupModel } from 'models/group-model';
 import { YubiKeyOtpModel } from 'models/otp-device/yubikey-otp-model';
 import { MenuModel } from 'models/menu/menu-model';
+import { AppModelBackupMixin } from 'models/app-model-backup';
+import { AppModelFileOpenMixin } from 'models/app-model-file-open';
+import { AppModelFilterMixin } from 'models/app-model-filter';
+import { AppModelSyncMixin } from 'models/app-model-sync';
 import { PluginManager } from 'plugins/plugin-manager';
 import { Features } from 'util/features';
-import { DateFormat } from 'comp/i18n/date-format';
 import { Launcher } from 'comp/launcher';
-import { UrlFormat } from 'util/formatting/url-format';
 import { IdGenerator } from 'util/generators/id-generator';
-import { Locale } from 'util/locale';
-import { StringFormat } from 'util/formatting/string-format';
-import { MenuItemModel } from 'models/menu/menu-item-model';
 import { Logger } from 'util/logger';
-import { noop } from 'util/fn';
-import debounce from 'lodash/debounce';
 import 'util/kdbxweb/protected-value-ex';
 
 class AppModel {
@@ -213,89 +205,6 @@ class AppModel {
         this.updateTags();
     }
 
-    _addTags(file) {
-        const tagsHash = {};
-        this.tags.forEach((tag) => {
-            tagsHash[tag.toLowerCase()] = true;
-        });
-        file.forEachEntry({}, (entry) => {
-            for (const tag of entry.tags) {
-                if (!tagsHash[tag.toLowerCase()]) {
-                    tagsHash[tag.toLowerCase()] = true;
-                    this.tags.push(tag);
-                }
-            }
-        });
-        this.tags.sort();
-    }
-
-    _tagsChanged() {
-        const section = this.menu.allItemsSection;
-        while (section.items.length > 1) {
-            section.items.pop();
-        }
-        if (this.tags.length) {
-            const tagItems = this.tags.map((tag) => {
-                return new MenuItemModel({
-                    title: tag,
-                    icon: 'tag',
-                    filterKey: 'tag',
-                    filterValue: tag,
-                    editable: true
-                });
-            });
-            section.addItem({
-                title: StringFormat.capFirst(Locale.tags),
-                icon: 'tags',
-                expanded: !!AppSettingsModel.tagsMenuExpanded || !!this.filter.tag,
-                items: tagItems,
-                sectionHeader: true,
-                persistExpandedKey: 'tagsMenuExpanded',
-                cls: 'menu__item--disclosure'
-            });
-        } else {
-            section.emit('change-items');
-        }
-    }
-
-    _presentGroupsMenu() {
-        const singleFile = this.files.length <= 1;
-        let hasVisibleSubgroups = false;
-        for (const root of this.menu.groupsSection.items) {
-            if (singleFile && root.top) {
-                root.cls = 'menu__item--hide-self';
-                if (!root.expanded) {
-                    root.expanded = true;
-                }
-                if (root.items) {
-                    hasVisibleSubgroups = root.items.some((group) => group.visible !== false);
-                }
-            } else {
-                root.cls = null;
-                hasVisibleSubgroups = true;
-            }
-        }
-        const showGroups = !singleFile || hasVisibleSubgroups;
-        this.menu.groupsSection.visible = showGroups;
-        this.menu.groupsSection.grow = showGroups;
-    }
-
-    updateTags() {
-        const oldTags = this.tags.slice();
-        this.tags.splice(0, this.tags.length);
-        for (const file of this.files) {
-            this._addTags(file);
-        }
-        if (oldTags.join(',') !== this.tags.join(',')) {
-            this._tagsChanged();
-        }
-    }
-
-    renameTag(from, to) {
-        this.files.forEach((file) => file.renameTag && file.renameTag(from, to));
-        this.updateTags();
-    }
-
     closeAllFiles() {
         if (!this.files.hasOpenFiles()) {
             return;
@@ -325,1096 +234,6 @@ class AppModel {
         this.menu.filesSection.removeByFile(file);
         this.menu.select({ item: this.menu.allItemsSection.items[0] });
         Events.emit('one-file-closed');
-    }
-
-    emptyTrash() {
-        this.files.forEach((file) => file.emptyTrash && file.emptyTrash());
-        this.refresh();
-    }
-
-    setFilter(filter) {
-        const keepAttachments =
-            filter.attachments === undefined && this.filter && this.filter.attachments;
-        this.filter = this.prepareFilter(filter);
-        if (keepAttachments) {
-            this.filter.attachments = true;
-        }
-        this.filter.subGroups = this.settings.expandGroups;
-        if (!this.filter.advanced && this.advancedSearch) {
-            this.filter.advanced = this.advancedSearch;
-        }
-        const entries = this.getEntries();
-        if (!this.activeEntryId || !entries.get(this.activeEntryId)) {
-            const firstEntry = entries[0];
-            this.activeEntryId = firstEntry ? firstEntry.id : null;
-        }
-        Events.emit('filter', { filter: this.filter, sort: this.sort, entries });
-        Events.emit('entry-selected', entries.get(this.activeEntryId));
-    }
-
-    refresh() {
-        this._presentGroupsMenu();
-        this.setFilter(this.filter);
-    }
-
-    selectEntry(entry) {
-        this.activeEntryId = entry.id;
-        this.refresh();
-    }
-
-    addFilter(filter) {
-        this.setFilter(Object.assign(this.filter, filter));
-    }
-
-    setSort(sort) {
-        this.sort = sort;
-        this.setFilter(this.filter);
-    }
-
-    getEntries() {
-        const entries = this.getEntriesByFilter(this.filter, this.files);
-        entries.sortEntries(this.sort, this.filter);
-        if (this.filter.trash) {
-            this.addTrashGroups(entries);
-        }
-        return entries;
-    }
-
-    getEntriesByFilter(filter, files) {
-        const preparedFilter = this.prepareFilter(filter);
-        const entries = new SearchResultCollection();
-
-        const devicesToMatchOtpEntries = files.filter((file) => file.backend === 'otp-device');
-
-        const matchedOtpEntrySet = this.settings.yubiKeyMatchEntries ? new Set() : undefined;
-
-        files
-            .filter((file) => file.backend !== 'otp-device')
-            .forEach((file) => {
-                file.forEachEntry(preparedFilter, (entry) => {
-                    if (matchedOtpEntrySet) {
-                        for (const device of devicesToMatchOtpEntries) {
-                            const matchingEntry = device.getMatchingEntry(entry);
-                            if (matchingEntry) {
-                                matchedOtpEntrySet.add(matchingEntry);
-                            }
-                        }
-                    }
-                    entries.push(entry);
-                });
-            });
-
-        if (devicesToMatchOtpEntries.length) {
-            for (const device of devicesToMatchOtpEntries) {
-                device.forEachEntry(preparedFilter, (entry) => {
-                    if (!matchedOtpEntrySet || !matchedOtpEntrySet.has(entry)) {
-                        entries.push(entry);
-                    }
-                });
-            }
-        }
-
-        return entries;
-    }
-
-    addTrashGroups(collection) {
-        this.files.forEach((file) => {
-            const trashGroup = file.getTrashGroup && file.getTrashGroup();
-            if (trashGroup) {
-                trashGroup.getOwnSubGroups().forEach((group) => {
-                    collection.unshift(GroupModel.fromGroup(group, file, trashGroup));
-                });
-            }
-        });
-    }
-
-    prepareFilter(filter) {
-        filter = { ...filter };
-
-        filter.textLower = filter.text ? filter.text.toLowerCase() : '';
-        filter.textParts = null;
-        filter.textLowerParts = null;
-
-        const exact = filter.advanced && filter.advanced.exact;
-        if (!exact && filter.text) {
-            const textParts = filter.text.split(/\s+/).filter((s) => s);
-            if (textParts.length) {
-                filter.textParts = textParts;
-                filter.textLowerParts = filter.textLower.split(/\s+/).filter((s) => s);
-            }
-        }
-
-        filter.tagLower = filter.tag ? filter.tag.toLowerCase() : '';
-
-        return filter;
-    }
-
-    getFirstSelectedGroupForCreation() {
-        const selGroupId = this.filter.group;
-        let file, group;
-        if (selGroupId) {
-            this.files.some((f) => {
-                file = f;
-                group = f.getGroup(selGroupId);
-                return group;
-            });
-        }
-        if (!group) {
-            file = this.files.find((f) => f.active && !f.readOnly);
-            group = file.groups[0];
-        }
-        return { group, file };
-    }
-
-    completeUserNames(part) {
-        const userNames = {};
-        this.files.forEach((file) => {
-            file.forEachEntry(
-                { text: part, textLower: part.toLowerCase(), advanced: { user: true } },
-                (entry) => {
-                    const userName = entry.user;
-                    if (userName) {
-                        userNames[userName] = (userNames[userName] || 0) + 1;
-                    }
-                }
-            );
-        });
-        const matches = Object.entries(userNames);
-        matches.sort((x, y) => y[1] - x[1]);
-        const maxResults = 5;
-        if (matches.length > maxResults) {
-            matches.length = maxResults;
-        }
-        return matches.map((m) => m[0]);
-    }
-
-    getEntryTemplates() {
-        const entryTemplates = [];
-        this.files.forEach((file) => {
-            file.forEachEntryTemplate?.((entry) => {
-                entryTemplates.push({ file, entry });
-            });
-        });
-        return entryTemplates;
-    }
-
-    canCreateEntries() {
-        return this.files.some((f) => f.active && !f.readOnly);
-    }
-
-    createNewEntry(args) {
-        const sel = this.getFirstSelectedGroupForCreation();
-        if (args?.template) {
-            if (sel.file !== args.template.file) {
-                sel.file = args.template.file;
-                sel.group = args.template.file.groups[0];
-            }
-            const templateEntry = args.template.entry;
-            const newEntry = EntryModel.newEntry(sel.group, sel.file);
-            newEntry.copyFromTemplate(templateEntry);
-            return newEntry;
-        } else {
-            return EntryModel.newEntry(sel.group, sel.file, {
-                tag: this.filter.tag
-            });
-        }
-    }
-
-    createNewEntryWithFields(group, fields) {
-        return EntryModel.newEntryWithFields(group, fields);
-    }
-
-    createNewGroup() {
-        const sel = this.getFirstSelectedGroupForCreation();
-        return GroupModel.newGroup(sel.group, sel.file);
-    }
-
-    createNewGroupWithName(group, file, name) {
-        const newGroup = GroupModel.newGroup(group, file);
-        newGroup.setName(name);
-        return newGroup;
-    }
-
-    createNewTemplateEntry() {
-        const file = this.getFirstSelectedGroupForCreation().file;
-        const group = file.getEntryTemplatesGroup() || file.createEntryTemplatesGroup();
-        return EntryModel.newEntry(group, file);
-    }
-
-    createDemoFile() {
-        if (!this.files.getByName('Demo')) {
-            const demoFile = new FileModel({ id: IdGenerator.uuid() });
-            demoFile.openDemo(() => {
-                this.addFile(demoFile);
-            });
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    createNewFile(name, callback) {
-        if (!name) {
-            for (let i = 0; ; i++) {
-                name = Locale.openNewFile + (i || '');
-                if (!this.files.getByName(name) && !this.fileInfos.getByName(name)) {
-                    break;
-                }
-            }
-        }
-        const newFile = new FileModel({ id: IdGenerator.uuid() });
-        newFile.create(name, () => {
-            this.addFile(newFile);
-            callback?.(newFile);
-        });
-    }
-
-    openFile(params, callback) {
-        const logger = new Logger('open', params.name);
-        logger.info('File open request');
-
-        const fileInfo = params.id
-            ? this.fileInfos.get(params.id)
-            : this.fileInfos.getMatch(params.storage, params.name, params.path);
-
-        if (!params.opts && fileInfo && fileInfo.opts) {
-            params.opts = fileInfo.opts;
-        }
-
-        if (fileInfo && fileInfo.modified) {
-            logger.info('Open file from cache because it is modified');
-            this.openFileFromCache(
-                params,
-                (err, file) => {
-                    if (!err && file) {
-                        logger.info('Sync just opened modified file');
-                        setTimeout(() => this.syncFile(file), 0);
-                    }
-                    callback(err);
-                },
-                fileInfo
-            );
-        } else if (params.fileData) {
-            logger.info('Open file from supplied content');
-            if (params.storage === 'file') {
-                Storage.file.stat(params.path, null, (err, stat) => {
-                    if (err) {
-                        return callback(err);
-                    }
-                    params.rev = stat.rev;
-                    this.openFileWithData(params, callback, fileInfo, params.fileData);
-                });
-            } else {
-                this.openFileWithData(params, callback, fileInfo, params.fileData, true);
-            }
-        } else if (!params.storage) {
-            logger.info('Open file from cache as main storage');
-            this.openFileFromCache(params, callback, fileInfo);
-        } else if (
-            fileInfo &&
-            fileInfo.openDate &&
-            fileInfo.rev === params.rev &&
-            fileInfo.storage !== 'file' &&
-            !this.settings.disableOfflineStorage
-        ) {
-            logger.info('Open file from cache because it is latest');
-            this.openFileFromCache(
-                params,
-                (err, file) => {
-                    if (err) {
-                        if (err.name === 'KdbxError' || err.ykError) {
-                            return callback(err);
-                        }
-                        logger.info(
-                            'Error loading file from cache, trying to open from storage',
-                            err
-                        );
-                        this.openFileFromStorage(params, callback, fileInfo, logger, true);
-                    } else {
-                        callback(err, file);
-                    }
-                },
-                fileInfo
-            );
-        } else if (
-            !fileInfo ||
-            !fileInfo.openDate ||
-            params.storage === 'file' ||
-            this.settings.disableOfflineStorage
-        ) {
-            this.openFileFromStorage(params, callback, fileInfo, logger);
-        } else {
-            logger.info('Open file from cache, will sync after load', params.storage);
-            this.openFileFromCache(
-                params,
-                (err, file) => {
-                    if (!err && file) {
-                        logger.info('Sync just opened file');
-                        setTimeout(() => this.syncFile(file), 0);
-                        callback(err);
-                    } else {
-                        if (err.name === 'KdbxError' || err.ykError) {
-                            return callback(err);
-                        }
-                        logger.info(
-                            'Error loading file from cache, trying to open from storage',
-                            err
-                        );
-                        this.openFileFromStorage(params, callback, fileInfo, logger, true);
-                    }
-                },
-                fileInfo
-            );
-        }
-    }
-
-    openFileFromCache(params, callback, fileInfo) {
-        Storage.cache.load(fileInfo.id, null, (err, data) => {
-            if (!data) {
-                err = Locale.openFileNoCacheError;
-            }
-            new Logger('open', params.name).info('Loaded file from cache', err);
-            if (err) {
-                callback(err);
-            } else {
-                this.openFileWithData(params, callback, fileInfo, data);
-            }
-        });
-    }
-
-    openFileFromStorage(params, callback, fileInfo, logger, noCache) {
-        logger.info('Open file from storage', params.storage);
-        const storage = Storage[params.storage];
-        const storageLoad = () => {
-            logger.info('Load from storage');
-            storage.load(params.path, params.opts, (err, data, stat) => {
-                if (err) {
-                    if (fileInfo && fileInfo.openDate && !this.settings.disableOfflineStorage) {
-                        logger.info('Open file from cache because of storage load error', err);
-                        this.openFileFromCache(params, callback, fileInfo);
-                    } else {
-                        logger.info('Storage load error', err);
-                        callback(err);
-                    }
-                } else {
-                    logger.info('Open file from content loaded from storage');
-                    params.fileData = data;
-                    params.rev = (stat && stat.rev) || null;
-                    const needSaveToCache = storage.name !== 'file';
-                    this.openFileWithData(params, callback, fileInfo, data, needSaveToCache);
-                }
-            });
-        };
-        const cacheRev = (fileInfo && fileInfo.rev) || null;
-        if (cacheRev && storage.stat) {
-            logger.info('Stat file');
-            storage.stat(params.path, params.opts, (err, stat) => {
-                if (
-                    !noCache &&
-                    fileInfo &&
-                    storage.name !== 'file' &&
-                    (err || (stat && stat.rev === cacheRev)) &&
-                    !this.settings.disableOfflineStorage
-                ) {
-                    logger.info(
-                        'Open file from cache because ' + (err ? 'stat error' : 'it is latest'),
-                        err
-                    );
-                    this.openFileFromCache(params, callback, fileInfo);
-                } else if (stat) {
-                    logger.info(
-                        'Open file from storage (' + stat.rev + ', local ' + cacheRev + ')'
-                    );
-                    storageLoad();
-                } else {
-                    logger.info('Stat error', err);
-                    callback(err);
-                }
-            });
-        } else {
-            storageLoad();
-        }
-    }
-
-    openFileWithData(params, callback, fileInfo, data, updateCacheOnSuccess) {
-        const logger = new Logger('open', params.name);
-        let needLoadKeyFile = false;
-        if (!params.keyFileData && fileInfo && fileInfo.keyFileName) {
-            params.keyFileName = fileInfo.keyFileName;
-            if (this.settings.rememberKeyFiles === 'data' && fileInfo.keyFileHash) {
-                params.keyFileData = FileModel.createKeyFileWithHash(fileInfo.keyFileHash);
-            } else if (this.settings.rememberKeyFiles === 'path' && fileInfo.keyFilePath) {
-                params.keyFilePath = fileInfo.keyFilePath;
-                if (Storage.file.enabled) {
-                    needLoadKeyFile = true;
-                }
-            }
-        } else if (params.keyFilePath && !params.keyFileData && !fileInfo) {
-            needLoadKeyFile = true;
-        }
-        const file = new FileModel({
-            id: fileInfo ? fileInfo.id : IdGenerator.uuid(),
-            name: params.name,
-            storage: params.storage,
-            path: params.path,
-            keyFileName: params.keyFileName,
-            keyFilePath: params.keyFilePath,
-            backup: fileInfo?.backup || null,
-            chalResp: params.chalResp
-        });
-        if (params.encryptedPassword) {
-            file.encryptedPassword = fileInfo.encryptedPassword;
-            file.encryptedPasswordDate = fileInfo?.encryptedPasswordDate || new Date();
-        }
-        const openComplete = (err) => {
-            if (err) {
-                return callback(err);
-            }
-            if (this.files.get(file.id)) {
-                return callback('Duplicate file id');
-            }
-            if (fileInfo && fileInfo.modified) {
-                if (fileInfo.editState) {
-                    logger.info('Loaded local edit state');
-                    file.setLocalEditState(fileInfo.editState);
-                }
-                logger.info('Mark file as modified');
-                file.modified = true;
-            }
-            if (fileInfo) {
-                file.syncDate = fileInfo.syncDate;
-            }
-            if (updateCacheOnSuccess && !this.settings.disableOfflineStorage) {
-                logger.info('Save loaded file to cache');
-                Storage.cache.save(file.id, null, params.fileData);
-            }
-            const rev = params.rev || (fileInfo && fileInfo.rev);
-            this.setFileOpts(file, params.opts);
-            this.addToLastOpenFiles(file, rev);
-            this.addFile(file);
-            callback(null, file);
-            this.fileOpened(file, data, params);
-        };
-        const open = () => {
-            file.open(params.password, data, params.keyFileData, openComplete);
-        };
-        if (needLoadKeyFile) {
-            Storage.file.load(params.keyFilePath, {}, (err, data) => {
-                if (err) {
-                    logger.info('Storage load error', err);
-                    callback(err);
-                } else {
-                    params.keyFileData = data;
-                    open();
-                }
-            });
-        } else {
-            open();
-        }
-    }
-
-    importFileWithXml(params, callback) {
-        const logger = new Logger('import', params.name);
-        logger.info('File import request with supplied xml');
-        const file = new FileModel({
-            id: IdGenerator.uuid(),
-            name: params.name,
-            storage: params.storage,
-            path: params.path
-        });
-        file.importWithXml(params.fileXml, (err) => {
-            logger.info('Import xml complete ' + (err ? 'with error' : ''), err);
-            if (err) {
-                return callback(err);
-            }
-            this.addFile(file);
-            this.fileOpened(file);
-        });
-    }
-
-    addToLastOpenFiles(file, rev) {
-        this.appLogger.debug(
-            'Add last open file',
-            file.id,
-            file.name,
-            file.storage,
-            file.path,
-            rev
-        );
-        const dt = new Date();
-        const fileInfo = new FileInfoModel({
-            id: file.id,
-            name: file.name,
-            storage: file.storage,
-            path: file.path,
-            opts: this.getStoreOpts(file),
-            modified: file.modified,
-            editState: file.getLocalEditState(),
-            rev,
-            syncDate: file.syncDate || dt,
-            openDate: dt,
-            backup: file.backup,
-            chalResp: file.chalResp
-        });
-        switch (this.settings.rememberKeyFiles) {
-            case 'data':
-                fileInfo.set({
-                    keyFileName: file.keyFileName || null,
-                    keyFileHash: file.getKeyFileHash()
-                });
-                break;
-            case 'path':
-                fileInfo.set({
-                    keyFileName: file.keyFileName || null,
-                    keyFilePath: file.keyFilePath || null
-                });
-        }
-        if (this.settings.deviceOwnerAuth === 'file' && file.encryptedPassword) {
-            const maxDate = new Date(file.encryptedPasswordDate);
-            maxDate.setMinutes(maxDate.getMinutes() + this.settings.deviceOwnerAuthTimeoutMinutes);
-            if (maxDate > new Date()) {
-                fileInfo.encryptedPassword = file.encryptedPassword;
-                fileInfo.encryptedPasswordDate = file.encryptedPasswordDate;
-            }
-        }
-        this.fileInfos.remove(file.id);
-        this.fileInfos.unshift(fileInfo);
-        this.fileInfos.save();
-    }
-
-    getStoreOpts(file) {
-        const opts = file.opts;
-        const storage = file.storage;
-        if (Storage[storage] && Storage[storage].fileOptsToStoreOpts && opts) {
-            return Storage[storage].fileOptsToStoreOpts(opts, file);
-        }
-        return null;
-    }
-
-    setFileOpts(file, opts) {
-        const storage = file.storage;
-        if (Storage[storage] && Storage[storage].storeOptsToFileOpts && opts) {
-            file.opts = Storage[storage].storeOptsToFileOpts(opts, file);
-        }
-    }
-
-    fileOpened(file, data, params) {
-        if (file.storage === 'file') {
-            Storage.file.watch(
-                file.path,
-                debounce(() => {
-                    this.syncFile(file);
-                }, Timeouts.FileChangeSync)
-            );
-        }
-        if (file.isKeyChangePending(true)) {
-            Events.emit('key-change-pending', { file });
-        }
-        const backup = file.backup;
-        if (data && backup && backup.enabled && backup.pending) {
-            this.scheduleBackupFile(file, data);
-        }
-        if (this.settings.yubiKeyAutoOpen) {
-            if (
-                this.attachedYubiKeysCount > 0 &&
-                !this.files.some((f) => f.backend === 'otp-device')
-            ) {
-                this.tryOpenOtpDeviceInBackground();
-            }
-        }
-        if (this.settings.deviceOwnerAuth) {
-            if (!params) {
-                return;
-            }
-            if (TouchIdRecovery.isRequired()) {
-                if (!params.encryptedPassword && TouchIdRecovery.isRequired(file.id)) {
-                    return TouchIdRecovery.recover(this, file, params, NativeModules)
-                        .then(() => this.appLogger.info('Touch ID reconnected'))
-                        .catch((error) => {
-                            this.appLogger.error('Error reconnecting Touch ID', error);
-                            TouchIdRecovery.showFailedAlert();
-                        });
-                }
-                return;
-            }
-            return this.saveEncryptedPassword(file, params);
-        }
-    }
-
-    fileClosed(file) {
-        if (file.storage === 'file') {
-            Storage.file.unwatch(file.path);
-        }
-    }
-
-    removeFileInfo(id) {
-        Storage.cache.remove(id);
-        this.fileInfos.remove(id);
-        this.fileInfos.save();
-    }
-
-    getFileInfo(file) {
-        return (
-            this.fileInfos.get(file.id) ||
-            this.fileInfos.getMatch(file.storage, file.name, file.path)
-        );
-    }
-
-    syncFile(file, options, callback) {
-        if (file.demo) {
-            return callback && callback();
-        }
-        if (file.syncing) {
-            return callback && callback('Sync in progress');
-        }
-        if (!file.active) {
-            return callback && callback('File is closed');
-        }
-        if (!options) {
-            options = {};
-        }
-        const logger = new Logger('sync', file.name);
-        const storage = options.storage || file.storage;
-        let path = options.path || file.path;
-        const opts = options.opts || file.opts;
-        if (storage && Storage[storage].getPathForName && (!path || storage !== file.storage)) {
-            path = Storage[storage].getPathForName(file.name);
-        }
-        const optionsForLogging = { ...options };
-        if (optionsForLogging.opts && optionsForLogging.opts.password) {
-            optionsForLogging.opts = { ...optionsForLogging.opts };
-            optionsForLogging.opts.password = '***';
-        }
-        logger.info('Sync started', storage, path, optionsForLogging);
-        let fileInfo = this.getFileInfo(file);
-        if (!fileInfo) {
-            logger.info('Create new file info');
-            const dt = new Date();
-            fileInfo = new FileInfoModel({
-                id: IdGenerator.uuid(),
-                name: file.name,
-                storage: file.storage,
-                path: file.path,
-                opts: this.getStoreOpts(file),
-                modified: file.modified,
-                editState: null,
-                rev: null,
-                syncDate: dt,
-                openDate: dt,
-                backup: file.backup
-            });
-        }
-        file.setSyncProgress();
-        const complete = (err) => {
-            if (!file.active) {
-                return callback && callback('File is closed');
-            }
-            logger.info('Sync finished', err || 'no error');
-            file.setSyncComplete(path, storage, err ? err.toString() : null);
-            fileInfo.set({
-                name: file.name,
-                storage,
-                path,
-                opts: this.getStoreOpts(file),
-                modified: file.dirty ? fileInfo.modified : file.modified,
-                editState: file.dirty ? fileInfo.editState : file.getLocalEditState(),
-                syncDate: file.syncDate,
-                chalResp: file.chalResp
-            });
-            if (this.settings.rememberKeyFiles === 'data') {
-                fileInfo.set({
-                    keyFileName: file.keyFileName || null,
-                    keyFileHash: file.getKeyFileHash()
-                });
-            }
-            if (!this.fileInfos.get(fileInfo.id)) {
-                this.fileInfos.unshift(fileInfo);
-            }
-            this.fileInfos.save();
-            if (callback) {
-                callback(err);
-            }
-        };
-        if (!storage) {
-            if (!file.modified && fileInfo.id === file.id) {
-                logger.info('Local, not modified');
-                return complete();
-            }
-            logger.info('Local, save to cache');
-            file.getData((data, err) => {
-                if (err) {
-                    return complete(err);
-                }
-                Storage.cache.save(fileInfo.id, null, data, (err) => {
-                    logger.info('Saved to cache', err || 'no error');
-                    complete(err);
-                    if (!err) {
-                        this.scheduleBackupFile(file, data);
-                    }
-                });
-            });
-        } else {
-            const maxLoadLoops = 3;
-            let loadLoops = 0;
-            const loadFromStorageAndMerge = () => {
-                if (++loadLoops === maxLoadLoops) {
-                    return complete('Too many load attempts');
-                }
-                logger.info('Load from storage, attempt ' + loadLoops);
-                Storage[storage].load(path, opts, (err, data, stat) => {
-                    logger.info('Load from storage', stat, err || 'no error');
-                    if (!file.active) {
-                        return complete('File is closed');
-                    }
-                    if (err) {
-                        return complete(err);
-                    }
-                    file.mergeOrUpdate(data, options.remoteKey, (err) => {
-                        logger.info('Merge complete', err || 'no error');
-                        this.refresh();
-                        if (err) {
-                            if (err.code === 'InvalidKey') {
-                                logger.info('Remote key changed, request to enter new key');
-                                Events.emit('remote-key-changed', { file });
-                            }
-                            return complete(err);
-                        }
-                        if (stat && stat.rev) {
-                            logger.info('Update rev in file info');
-                            fileInfo.rev = stat.rev;
-                        }
-                        file.syncDate = new Date();
-                        if (file.modified) {
-                            logger.info('Updated sync date, saving modified file');
-                            saveToCacheAndStorage();
-                        } else if (file.dirty) {
-                            if (this.settings.disableOfflineStorage) {
-                                logger.info('File is dirty and cache is disabled');
-                                return complete(err);
-                            }
-                            logger.info('Saving not modified dirty file to cache');
-                            Storage.cache.save(fileInfo.id, null, data, (err) => {
-                                if (err) {
-                                    return complete(err);
-                                }
-                                file.dirty = false;
-                                logger.info('Complete, remove dirty flag');
-                                complete();
-                            });
-                        } else {
-                            logger.info('Complete, no changes');
-                            complete();
-                        }
-                    });
-                });
-            };
-            const saveToStorage = (data) => {
-                logger.info('Save data to storage');
-                const storageRev = fileInfo.storage === storage ? fileInfo.rev : undefined;
-                Storage[storage].save(
-                    path,
-                    opts,
-                    data,
-                    (err, stat) => {
-                        if (err && err.revConflict) {
-                            logger.info('Save rev conflict, reloading from storage');
-                            loadFromStorageAndMerge();
-                        } else if (err) {
-                            logger.info('Error saving data to storage');
-                            complete(err);
-                        } else {
-                            if (stat && stat.rev) {
-                                logger.info('Update rev in file info');
-                                fileInfo.rev = stat.rev;
-                            }
-                            if (stat && stat.path) {
-                                logger.info('Update path in file info', stat.path);
-                                file.path = stat.path;
-                                fileInfo.path = stat.path;
-                                path = stat.path;
-                            }
-                            file.syncDate = new Date();
-                            logger.info('Save to storage complete, update sync date');
-                            this.scheduleBackupFile(file, data);
-                            complete();
-                        }
-                    },
-                    storageRev
-                );
-            };
-            const saveToCacheAndStorage = () => {
-                logger.info('Getting file data for saving');
-                file.getData((data, err) => {
-                    if (err) {
-                        return complete(err);
-                    }
-                    if (storage === 'file') {
-                        logger.info('Saving to file storage');
-                        saveToStorage(data);
-                    } else if (!file.dirty) {
-                        logger.info('Saving to storage, skip cache because not dirty');
-                        saveToStorage(data);
-                    } else if (this.settings.disableOfflineStorage) {
-                        logger.info('Saving to storage because cache is disabled');
-                        saveToStorage(data);
-                    } else {
-                        logger.info('Saving to cache');
-                        Storage.cache.save(fileInfo.id, null, data, (err) => {
-                            if (err) {
-                                return complete(err);
-                            }
-                            file.dirty = false;
-                            logger.info('Saved to cache, saving to storage');
-                            saveToStorage(data);
-                        });
-                    }
-                });
-            };
-            logger.info('Stat file');
-            Storage[storage].stat(path, opts, (err, stat) => {
-                if (!file.active) {
-                    return complete('File is closed');
-                }
-                if (err) {
-                    if (err.notFound) {
-                        logger.info('File does not exist in storage, creating');
-                        saveToCacheAndStorage();
-                    } else if (file.dirty) {
-                        if (this.settings.disableOfflineStorage) {
-                            logger.info('Stat error, dirty, cache is disabled', err || 'no error');
-                            return complete(err);
-                        }
-                        logger.info('Stat error, dirty, save to cache', err || 'no error');
-                        file.getData((data, e) => {
-                            if (e) {
-                                logger.error('Error getting file data', e);
-                                return complete(err);
-                            }
-                            Storage.cache.save(fileInfo.id, null, data, (e) => {
-                                if (e) {
-                                    logger.error('Error saving to cache', e);
-                                }
-                                if (!e) {
-                                    file.dirty = false;
-                                }
-                                logger.info('Saved to cache, exit with error', err || 'no error');
-                                complete(err);
-                            });
-                        });
-                    } else {
-                        logger.info('Stat error, not dirty', err || 'no error');
-                        complete(err);
-                    }
-                } else if (stat.rev === fileInfo.rev) {
-                    if (file.modified) {
-                        logger.info('Stat found same version, modified, saving');
-                        saveToCacheAndStorage();
-                    } else {
-                        logger.info('Stat found same version, not modified');
-                        complete();
-                    }
-                } else {
-                    logger.info('Found new version, loading from storage');
-                    loadFromStorageAndMerge();
-                }
-            });
-        }
-    }
-
-    deleteAllCachedFiles() {
-        for (const fileInfo of this.fileInfos) {
-            if (fileInfo.storage && !fileInfo.modified) {
-                Storage.cache.remove(fileInfo.id);
-            }
-        }
-    }
-
-    clearStoredKeyFiles() {
-        for (const fileInfo of this.fileInfos) {
-            fileInfo.set({
-                keyFileName: null,
-                keyFilePath: null,
-                keyFileHash: null
-            });
-        }
-        this.fileInfos.save();
-    }
-
-    unsetKeyFile(fileId) {
-        const fileInfo = this.fileInfos.get(fileId);
-        fileInfo.set({
-            keyFileName: null,
-            keyFilePath: null,
-            keyFileHash: null
-        });
-        this.fileInfos.save();
-    }
-
-    setFileBackup(fileId, backup) {
-        const fileInfo = this.fileInfos.get(fileId);
-        if (fileInfo) {
-            fileInfo.backup = backup;
-        }
-        this.fileInfos.save();
-    }
-
-    _ensureDefaultBackup(file) {
-        let defaultPath = this.settings.backupDefaultPath;
-        if (!Launcher || !defaultPath || file.backup || file.storage !== 'file') {
-            return;
-        }
-        defaultPath = Launcher.expandHomePath(defaultPath);
-        const backup = {
-            enabled: true,
-            storage: 'file',
-            path: defaultPath + '/' + file.name + '.kdbx.{date}.bak',
-            schedule: '1d',
-            pending: true
-        };
-        file.backup = backup;
-        this.setFileBackup(file.id, backup);
-    }
-
-    pruneBackups(backup, logger) {
-        const maxCount = this.settings.backupMaxCount;
-        if (!Launcher || !maxCount || backup.storage !== 'file') {
-            return;
-        }
-        const sep = backup.path.lastIndexOf('/');
-        const folder = backup.path.substring(0, sep);
-        const namePattern = backup.path.substring(sep + 1);
-        const dateIx = namePattern.indexOf('{date}');
-        if (dateIx < 0) {
-            return;
-        }
-        const prefix = namePattern.substring(0, dateIx);
-        const suffix = namePattern.substring(dateIx + '{date}'.length);
-        Launcher.listDir(folder, (err, entries) => {
-            if (err || !entries) {
-                return;
-            }
-            const backups = entries
-                .filter(
-                    (name) =>
-                        name.length > prefix.length + suffix.length &&
-                        name.startsWith(prefix) &&
-                        name.endsWith(suffix)
-                )
-                .sort();
-            for (const name of backups.slice(0, -maxCount)) {
-                logger.info('Pruning old backup', name);
-                Launcher.deleteFile(folder + '/' + name);
-            }
-        });
-    }
-
-    backupFile(file, data, callback) {
-        const opts = file.opts;
-        let backup = file.backup;
-        const logger = new Logger('backup', file.name);
-        if (!backup || !backup.storage || !backup.path) {
-            return callback('Invalid backup settings');
-        }
-        let path = backup.path.replace('{date}', DateFormat.dtStrFs(new Date()));
-        logger.info('Backup file to', backup.storage, path);
-        const saveToFolder = () => {
-            if (Storage[backup.storage].getPathForName) {
-                path = Storage[backup.storage].getPathForName(path);
-            }
-            Storage[backup.storage].save(path, opts, data, (err) => {
-                if (err) {
-                    logger.error('Backup error', err);
-                } else {
-                    logger.info('Backup complete');
-                    backup = file.backup;
-                    backup.lastTime = Date.now();
-                    delete backup.pending;
-                    file.backup = backup;
-                    this.setFileBackup(file.id, backup);
-                    this.pruneBackups(backup, logger);
-                }
-                callback(err);
-            });
-        };
-        let folderPath = UrlFormat.fileToDir(path);
-        if (Storage[backup.storage].getPathForName) {
-            folderPath = Storage[backup.storage].getPathForName(folderPath).replace('.kdbx', '');
-        }
-        Storage[backup.storage].stat(folderPath, opts, (err) => {
-            if (err) {
-                if (err.notFound) {
-                    logger.info('Backup folder does not exist');
-                    if (!Storage[backup.storage].mkdir) {
-                        return callback('Mkdir not supported by ' + backup.storage);
-                    }
-                    Storage[backup.storage].mkdir(folderPath, (err) => {
-                        if (err) {
-                            logger.error('Error creating backup folder', err);
-                            callback('Error creating backup folder');
-                        } else {
-                            logger.info('Backup folder created');
-                            saveToFolder();
-                        }
-                    });
-                } else {
-                    logger.error('Stat folder error', err);
-                    callback('Cannot stat backup folder');
-                }
-            } else {
-                logger.info('Backup folder exists, saving');
-                saveToFolder();
-            }
-        });
-    }
-
-    scheduleBackupFile(file, data) {
-        const backup = file.backup;
-        if (!backup || !backup.enabled) {
-            return;
-        }
-        const logger = new Logger('backup', file.name);
-        let needBackup = false;
-        if (!backup.lastTime) {
-            needBackup = true;
-            logger.debug('No last backup time, backup now');
-        } else {
-            const dt = new Date(backup.lastTime);
-            switch (backup.schedule) {
-                case '0':
-                    break;
-                case '1d':
-                    dt.setDate(dt.getDate() + 1);
-                    break;
-                case '1w':
-                    dt.setDate(dt.getDate() + 7);
-                    break;
-                case '1m':
-                    dt.setMonth(dt.getMonth() + 1);
-                    break;
-                default:
-                    return;
-            }
-            if (dt.getTime() <= Date.now()) {
-                needBackup = true;
-            }
-            logger.debug(
-                'Last backup time: ' +
-                    new Date(backup.lastTime) +
-                    ', schedule: ' +
-                    backup.schedule +
-                    ', next time: ' +
-                    dt +
-                    ', ' +
-                    (needBackup ? 'backup now' : 'skip backup')
-            );
-        }
-        if (!backup.pending) {
-            backup.pending = true;
-            this.setFileBackup(file.id, backup);
-        }
-        if (needBackup) {
-            this.backupFile(file, data, noop);
-        }
     }
 
     usbDevicesChanged() {
@@ -1604,5 +423,10 @@ class AppModel {
         }
     }
 }
+
+Object.assign(AppModel.prototype, AppModelFilterMixin);
+Object.assign(AppModel.prototype, AppModelFileOpenMixin);
+Object.assign(AppModel.prototype, AppModelSyncMixin);
+Object.assign(AppModel.prototype, AppModelBackupMixin);
 
 export { AppModel };
